+++ Constraid/FlushWithMargins.swift
@@ -4,74 +4,95 @@
 import UIKit
 
 extension ConstraidView {
+    @discardableResult
     open func flush(withLeadingMarginOf item: Any?, constant: CGFloat = 0.0,
-                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) {
+                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
 
         self.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
+        let constraints = ConstraidConstraintCollection([
             NSLayoutConstraint(item: self, attribute: .leading, relatedBy: .equal,
                                toItem: item, attribute: .leadingMargin, multiplier: multiplier,
                                constant: constant, priority: priority)
             ])
+        constraints.activate()
+        return constraints
     }
 
+    @discardableResult
     open func flush(withTrailingMarginOf item: Any?, constant: CGFloat = 0.0,
-                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) {
+                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
 
         self.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
+        let constraints = ConstraidConstraintCollection([
             NSLayoutConstraint(item: self, attribute: .trailing, relatedBy: .equal,
                                toItem: item, attribute: .trailingMargin, multiplier: multiplier,
                                constant: constant, priority: priority)
             ])
+        constraints.activate()
+        return constraints
     }
 
+    @discardableResult
     open func flush(withTopMarginOf item: Any?, constant: CGFloat = 0.0,
-                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) {
+                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
 
         self.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
+        let constraints = ConstraidConstraintCollection([
             NSLayoutConstraint(item: self, attribute: .top, relatedBy: .equal,
                                toItem: item, attribute: .topMargin, multiplier: multiplier,
                                constant: constant, priority: priority)
             ])
+        constraints.activate()
+        return constraints
     }
 
+    @discardableResult
     open func flush(withBottomMarginOf item: Any?, constant: CGFloat = 0.0,
-                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) {
+                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
 
         self.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
+        let constraints = ConstraidConstraintCollection([
             NSLayoutConstraint(item: self, attribute: .bottom, relatedBy: .equal,
                                toItem: item, attribute: .bottomMargin, multiplier: multiplier,
                                constant: constant, priority: priority)
             ])
+        constraints.activate()
+        return constraints
     }
 
+    @discardableResult
     open func flush(withVerticalMarginsOf item: Any?, constant: CGFloat = 0.0,
-                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) {
+                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
 
-        flush(withLeadingMarginOf: item, constant: constant, multiplier: multiplier,
-              priority: priority)
-        flush(withTrailingMarginOf: item, constant: constant, multiplier: multiplier,
-              priority: priority)
+        let constraints = flush(withLeadingMarginOf: item, constant: constant,
+                                multiplier: multiplier, priority: priority) +
+                          flush(withTrailingMarginOf: item, constant: (-1.0 * constant),
+                                multiplier: multiplier, priority: priority)
+        constraints.activate()
+        return constraints
     }
 
+    @discardableResult
     open func flush(withHorizontalMarginsOf item: Any?, constant: CGFloat = 0.0,
-                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) {
+                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
 
-        flush(withTopMarginOf: item, constant: constant, multiplier: multiplier,
-              priority: priority)
-        flush(withBottomMarginOf: item, constant: constant, multiplier: multiplier,
-              priority: priority)
+        let constraints = flush(withTopMarginOf: item, constant: constant, multiplier: multiplier,
+                                priority: priority) +
+                          flush(withBottomMarginOf: item, constant: (-1.0 * constant),
+                                multiplier: multiplier, priority: priority)
+        constraints.activate()
+        return constraints
     }
 
+    @discardableResult
     open func flush(withMarginsOf item: Any?, constant: CGFloat = 0.0,
-                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) {
+                    multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
 
-        flush(withHorizontalMarginsOf: item, constant: constant, multiplier: multiplier,
-              priority: priority)
-        flush(withVerticalMarginsOf: item, constant: constant, multiplier: multiplier,
-              priority: priority)
+        let constraints = flush(withHorizontalMarginsOf: item, constant: constant,
+                                multiplier: multiplier, priority: priority) +
+                          flush(withVerticalMarginsOf: item, constant: constant,
+                                multiplier: multiplier, priority: priority)
+        constraints.activate()
+        return constraints
     }
 }