+++ CHANGELOG.md
@@ -10,6 +10,9 @@ that you can set version constraints properly.
 
 #### [Unreleased][unreleased]
 
+* `Changed`: `flush(withMargin...)` methods to return collection of constraints
+* `Fixed`: issue where flush(withMarginsOf), flush(withVerticalMarginsOf), and
+  flush(withHorizontalMarginsOf) constant argument would not do what expected
 * `Changed`: `flush(withEdge...)` methods to return collection of constraints
 * `Fixed`: issue where flush(withEdgesOf), flush(withVerticalEdgesOf), and
   flush(withHorizontalEdgesOf) constant argument would not do what expected